+++ src/QGCApplication.h
@@ -42,6 +42,7 @@
 class QGCSingleton;
 class MainWindow;
 class QGCToolbox;
+class QGCFileDownload;
 
 /**
  * @brief The main application and management class.
@@ -152,9 +153,13 @@ public:
 
 private slots:
     void _missingParamsDisplay(void);
+    void _currentVersionDownloadFinished(QString remoteFile, QString localFile);
+    void _currentVersionDownloadError(QString errorMsg);
+    bool _parseVersionText(const QString& versionString, int& majorVersion, int& minorVersion, int& buildVersion);
 
 private:
     QObject* _rootQmlObject(void);
+    void _checkForNewVersion(void);
 
 #ifdef __mobile__
     QQmlApplicationEngine* _qmlAppEngine;
@@ -170,6 +175,10 @@ private:
     QStringList         _missingParams;                                     ///< List of missing facts to be displayed
     bool				_fakeMobile;                                        ///< true: Fake ui into displaying mobile interface
     bool                _settingsUpgraded;                                  ///< true: Settings format has been upgrade to new version
+    int                 _majorVersion;
+    int                 _minorVersion;
+    int                 _buildVersion;
+    QGCFileDownload*    _currentVersionDownload;
 
 #ifdef QT_DEBUG
     bool _testHighDPI;  ///< true: double fonts sizes for simulating high dpi devices