+++ ChangeLog.md
@@ -11,6 +11,7 @@ Note: This file only contains high level features or important fixes.
 ## 3.4
 
 ### 3.4.4 - Not yet released
+* Multi-Vehicle Start Mission and Pause now work correctly. Issue #6864.
 
 ### 3.4.3
 * Fix bug where Resume Mission would not display correctly in some cases. Issue #6835.
@@ -22,3 +23,4 @@ Note: This file only contains high level features or important fixes.
 ### 3.4.1
 * Fix crash when Survery with terrain follow is moved quickly
 * Fix terrain follow climb/descent rate fields swapped in ui
+